[lib] Ignore exceptions in extractUserIDsFromPayload
Summary:
When an endpoint fails, it will respond with something that doesn't match the expected output validator.

This causes the client to throw a different error that the one we expect from the endpoint failing.

This diff addresses [ENG-8324](https://linear.app/comm/issue/ENG-8324/all-invite-link-errors-show-up-as-unknown-error) by wrapping the `convertObject` call in a `try-catch` that ignores all exceptions.

Test Plan:
I tested on both `native` and `web` by attempting to set an invite link corresponding to a reserved string. Before this diff I got a generic error, but afterwards I got the expected, more precise error

Reviewers: inka, varun, will

Subscribers: