[native] Ignore edit button presses while saving
Summary:
If you're editing an `Entry`, change the text, and then hit the "Save" button, the first event that fires now is `onBlur` (see [here](https://linear.app/comm/issue/ENG-8325/pressing-s
ave-entry-button-while-editing-on-native-leads-to-re-focus#comment-1a9d77c2) for more context).

The `onBlur` results in a save being triggered, and `this.state.editing` is unset. After that, the save button `onPress` is now triggered (previously this came first).

To avoid the `onPress` causing us to go back into an editing mode, we can ignore it here while the save is ongoing.

Test Plan: Confirm that we don't refocus the `Entry` if you hit "Save" after editing and changing the text

Reviewers: tomek

Subscribers: