[native] Apply pending delete before dispatching RawEntryInfo updates
Summary:
If you're editing an `Entry`, change the text, and then hit the "Delete" button, the first event that fires now is `onBlur` (see [here](https://linear.app/comm/issue/ENG-8325/pressing-save-entry-button-while-editing-on-native-leads-to-re-focus#comment-1a9d77c2) for more context).

This means that we first attempt to save the updated text, and then we attempt to delete the `Entry`. This leads to an issue: we optimistically set `deleted: true` for the `Entry`, but when the save comes in it flips that to `false`, resulting in a weird behavior in the UI.

This diff avoids that by patching the result from the save to have `deleted: true`.

Depends on D12260

Test Plan: Make sure that the `Entry` doesn't "rise from the dead" after disappearing when you hit "Delete" after editing and changing the text

Reviewers: tomek

Differential Revision: https://phab.comm.dev/D12261