[native] Fix Entry deletion after creation
Summary:
Noticed that `this.needsDeleteAfterCreation` wasn't working as intended. The issue was that we were getting a `concurrent_modification` error because the `prevText` we were passing in for the delete did not match the text after creation. We were passing in the empty string since that is what `this.props.entryInfo.text` was prior to the `CREATE_ENTRY_SUCCESS` action being reduced.

To address this, this diff allows us to pass `this.state.text` in instead. This makes sure that we match what is in the keyserver MariaDB when we attempt the deletion.

Depends on D12262

Test Plan: Type an entry and then press the delete button. On web, confirm that it appears in the history, but it otherwise missing

Reviewers: tomek

Subscribers: