[terraform] set up self-host terraform with vpc
Summary: This sets up the initial self-host configuration along with a vpc. The vpc includes a public and private subnet. I plan to have the keyserver hosted in the public subnet and all other required services (i.e. MariaDB) stored in the private.

Test Plan: Tested on personal AWS. `terraform apply` was successful

Reviewers: varun, bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12277