[lib] Replace `threadMemberHasPermission` logic in `appendUserInfo`
Summary:
Instead of checking `KNOW_OF` using `threadMemberHasPermission` which requires `member.permissions` for all members, we use the member's role permissions to determine if they have `KNOW_OF`.

This was the one usage of `threadMemberHasPermission`, so now it can be fully removed as well.

---

Depends on D12158

Test Plan: Added log statements at each step and ensured that values were as expected:

Reviewers: ashoat!, tomek, varun, ginsu

Differential Revision: https://phab.comm.dev/D12259