[identity] Add Serialize trait for inbound keys types
Summary:
Added serialize derive for structs that are part of the inbound keys response
This will be required for sending these in the inbound keys HTTP response

Alternatively, I could create a dedicated JSON-only structs and provide conversion traits, not to obfuscate existing structs, but this can be done at any time as a follow up.

Test Plan:
In `get_inbound_keys` handler, used `serde_json::to_string()` to print out results.
The JSON structure is as proposed in [[ https://linear.app/comm/issue/ENG-8290/call-the-http-endpoint-to-fetch-inbound-keys-from-the-nse#comment-c6deb0c3 | ENG-8290 ]].

Reviewers: marcin!

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12300