[blob] Don't check if the invite link is reserved
Summary:
We're checking this only on the keyserver side for now.

https://linear.app/comm/issue/ENG-8302/stop-checking-if-an-invite-link-contains-a-reserved-username-on-blob

Test Plan: Run a local blob service, comment out reserved link check on a keyserver. Create a reserved link, open it and verify that the blob was downloaded using the browser inspector. Also check if offensive links check still works on keyserver.

Reviewers: bartek, varun