[lib] add reserved username workflow to identity login actions
Summary:
in our identity login actions, we should check the exception message and if it tells us that the username is reserved, we should attempt reserved registration (if possible, i.e. if on native)

Depends on D11911

Test Plan: on staging, i set the keyserver public key env var to my local keyserver's device ID. then i registered a user directly with my keyserver. flipped usingCSAT to true and tried logging in again. confirmed that i was registered with the identity service and granted an access token

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D11912