[native] Fetch invite links when entering thread settings
Summary:
part of [[https://linear.app/comm/issue/ENG-7793/disallow-adding-users-to-community-roots-on-the-client | ENG-7793]]
while testing I noticed that sometimes when I enter thread settings, a thread that does have an invite link appears to not have it. This is because we used to only fetch invite links when opening the community drawer, because this was the only place we were displaying them. Now I am also displaying them in thread settings, and the user may navigate to them without ever opening the drawer.
This is not the best solution, but this is a compromise we decided on when implementing invite links to avoid implementing updates for them

Test Plan: Tested that now when the user opens the app and navigates to thread settings immediately, the invite link is fetched

Reviewers: tomek, kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12286