[lib] Rename `MemberInfo` to `MemberInfoWithPermissions`
Summary:
This is to make room for `MemberInfoWithoutPermissions` or `MemberInfoSansPermissions, etc.

We're going to need both in different places in the codebase. Specifically we'll definitely need `MemberInfoWithPermissions` when constructing `RawThreadInfo`s for old clients, etc.

The goal at the end of the stack is for the client to only deal with `MemberInfoSansPermissions`.

Will try to make the changes as minimal as possible...

---

Depends on D12295

Test Plan: `flow`/NA just a simple rename

Reviewers: ashoat, ginsu, tomek, varun

Differential Revision: https://phab.comm.dev/D12296