[keyserver][lib] Add ACKNOWLEDGE relationship action
Summary:
issue: [[https://linear.app/comm/issue/ENG-7822/create-a-keyserver-endpoint-for-fetching-userinfos | ENG-7822]]
We need to fetch avatars for users the client has become aware of. We want the keyserver to remember that the users now know of each other and send an update to the second user.
So we will use the updates to obtain the avatar - we will call an endpoint that will update the relationshipt to "KNOW_OF", if there is no other undirected relationship. This endpoint will create updates if needed

Test Plan: Dispatched `updateRelationship` with `relationshipActions.ACKNOWLEDGE` and checked that both users get and update, and if the user has an avatar set, this avatar is sent in the update. Checked in the keyserver db that the users now have a `KNOW_OF` relationship

Reviewers: tomek, kamil

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12313