[native] Open a community when a user clicks an invite link and is already a member
Summary:
Check if a user is already a member and navigate to a thread.

{F1955129}

https://linear.app/comm/issue/ENG-8162/open-a-community-when-a-user-clicks-a-link-and-is-already-a-member

Depends on D12316

Test Plan:
Click a link to a community where a user is a member - check if a modal is displayed and after closing it a proper thread is displayed.
Click a link to a community where a user isn't a member - the old experience should be shown.

Reviewers: inka, ginsu

Reviewed By: inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12317