[lib][web][native] Navigate to a thread after joining a community
Summary: Instead of closing a modal immediately, we schedule a new effect that will navigate to a thread or close a modal, depending on the presence of the thread in the store. We need to run this as an effect in the next render, because we need to wait for the successful join action to be dispatched.

https://linear.app/comm/issue/ENG-8018/navigate-to-a-community-after-accepting-an-invite-link

Depends on D12317

Test Plan: On both web and native: check if the behavior is correct when a user joins a community, is already a member, or a link is invalid.

Reviewers: inka, ginsu