[lib] Don't remove avatar when fetching username
Summary:
When `findUserIdentitiesActionTypes.success` is dispatched, an entry for this id may already exist in the store - we may have gotten the user infos from auth keyserver already. We don't want to forget other values of this user info. Namely - `avatar` and `relationshipStatus`.

Test Plan:
Dispatched `findUserIdentitiesActionTypes.success` without avatar for an entry that existed in the store and had an avatar. Checked that the updated entry still had the avatar.

Reviewers:
tomek, kamil
Subscribers: