[lib] Fetch user avatars for new user ids
Summary:
issue: [[https://linear.app/comm/issue/ENG-7823/call-keyserver-endpoint-to-fetch-new-userinfos | ENG-7823]]
After calling update relationships endpoint, the keyserver will generate updates and send them to the client. This way we will obtain the `avatar`.
For now, this action will override the username for this id if already present in the store. This will be changed in [[https://linear.app/comm/issue/ENG-8365/stop-saving-usernames-from-auth-keyserver | ENG-8365]].

Test Plan:
Dispatched `processNewUserIDsActionType` with user id that the current user didn't have in the store. Checked that user infos are fetched from the keyserver and saved in the store.

Reviewers:
tomek, kamil
Subscribers: