[keyserver][lib] Introduce fetch_pending_updates endpoint
Summary:
This addresses [ENG-8088](https://linear.app/comm/issue/ENG-8088/introduce-new-keyserver-rest-endpoint-for-healing-from-a-socket-crash). In a later diff, the client will call this when encountering a socket crash loop. The higher timeout will allow us to download a larger amount of data.

Depends on D12363

Test Plan:
In combination with later diffs, I tested as follows:

1. I created a socket crash loop on a physical iOS device using @inka's `create-many-threads-to-trigger-crash-loop.js` script (see [ENG-8090](https://linear.app/comm/issue/ENG-8090/test-to-confirm-we-can-heal-from-a-socket-crash-loop)
2. I confirmed that the socket was unable to connect prior to my diff stack
3. I confirmed that after applying my diff stack, the `SUCCESS` action was dispatched, and the socket was able to connect afterwards

Reviewers: tomek, kamil

Reviewed By: tomek, kamil

Subscribers: kamil, inka

Differential Revision: https://phab.comm.dev/D12364