[keyserver] Bump serverResponseTimeout to 10s
Summary:
`serverResponseTimeout` is used only in `keyserver/src/socket/socket.js`, and is the time before the keyserver will time out a Websocket connection with a client.

This change brings it in line with `clientRequestSocketTimeout`, which is the client's timeout for a Websocket connection to the keyserver.

I've had this change for some time on my production keyserver and never diffed it up for some reason.

Test Plan: Been running it in production for some time

Reviewers: will, tomek

Reviewed By: will

Differential Revision: https://phab.comm.dev/D12338