[lib] Discard keyserver usernames for joinThreadActionTypes and newThreadActionTypes
Summary:
issue: [[https://linear.app/comm/issue/ENG-8365/stop-saving-usernames-from-auth-keyserver | ENG-8365]]
We want the username to be fetched from identity. This is handled by UserInfosHandler, which checks if there are any user infos with null username

Test Plan:
Joined a community, checked that user infos for new users had a `null` id, and for user the user already knew the username was unchanged.
Checked that after a while the usernames were fetched from identity

Reviewers: tomek, kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12352