[web] introduce component to ping identity
Summary: we want to introduce this component now so that we can check in the next web release if the production identity service is reachable. before this change, ashoat's keyserver was the only device that called identity. going forward, all web clients should be connecting to identity as well.

Test Plan: saw "Identity ping successful" log in browser console

Reviewers: ashoat!

Subscribers: