[lib] Set new session if received from fetch_pending_updates
Summary:
The logic in the keyserver's `fetchDataForSocketInit` method is written to handle session setting in a peculiar way. The normal session setting mechanisms rely on the result being processed by `callSingleKeyserverEndpoint`. When a session is set during a socket initialization, the new session is written into the result, and the caller is expected to extract it and handle it manually.

Since `fetch_pending_updates` is called from `callSingleKeyserverEndpoint` on the client, we could consider an alternative here, where `fetchDataForSocketInit` would only behave this way when called from an actual socket (as opposed to `fetch_pending_updates`). I decided it would be more simple to leave the keyserver code as-is, and to handle the received session on the client side. Open to the alternative if it's preferred.

Depends on D12387

Test Plan:
Prior to this, my test plan in D12386 would repopulate some data, but not all data. The issue is that we were treating it as an `INCREMENTAL` rather than `FULL`, but `INCREMENTAL` was not possible from time 0 because some of the updates since that point had been deleted.

In additional to that, I did some work to verify the safety of this change:

1. I tested registration via the new registration flow, and confirmed that nothing weird happened (this is the other case where `serverTime` is set to 0)
2. I looked through the login/registration code on keyserver and on the client to confirm that there were only two cases where `serverTime` was set to 0

Reviewers: kamil

Subscribers: