Move all notif types to lib except for APNs notifs
Summary: This differential moves all notif types to lib except for APNs. The reason for this is that APNs types are provided by `node-apn` that can't be used in lib. Therefore APNs notifs types, generation and encryption will be handled in a separate diff since it is less copy-pasting and more implementing from scratch.

Test Plan:
1. Flow
2. Test all notif types for all platforms and large notifs as well.

Reviewers: kamil, tomek

Reviewed By: kamil, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12394