[keyserver] Extend link validation to support thread links
Summary:
In addition to the previous solution, fetch thread-related data and return the response based on both community and thread data.

https://linear.app/comm/issue/ENG-8249/modify-verify-invite-links-response-to-include-a-thread-id-and-its

Depends on D12354

Test Plan:
Checked returned verification status for community and thread links when viewer is and isn't a member.
Also checked if joining a thread link results in joining a community.

These two checks additionally validate that the older client will work correctly with the new solution to links.

Reviewers: inka, ginsu

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12392