[native][web] Navigate to a thread if already joined
Summary:
When a user already joined a thread or a community, check the most specific chat (a thread is more specific than its community because the fact that a user is a member of a thread requires the user to be a member of this thread's community) and navigate to it if appropriate.

https://linear.app/comm/issue/ENG-8251/navigate-to-a-community-or-a-thread-if-present-in-the-response

Depends on D12399

Test Plan: Tested on both native and web. Used a link to a thread and to a community where the viewer was a member and verified that a thread or a community was opened.

Reviewers: inka, ginsu