[native][web][lib] add versionSupported method to IdentityServiceClient
Summary:
Depends on 12404

Initially, we just needed to call versionSupported on native. Now, we want to call it on both web and native. Before that, we should add this method to the IdentityServiceClient so we can call it from lib/actions/user-actions.js (next diff)

Test Plan: tested in next diff by dispatching action on native that calls this method

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D12405