Bump minimal native code version to 149
Summary: I want to remove support for 135 code version not to have to handle `managedAps: {}` in APNs rescind types I am about to introduce. The next code version after 135 that we have checks for is 150. Therefore I cut support for less than 149 to minimize the amout of checks to remove and the scope of this diff. The oldest code version in use is currently 185 (1 ios user).

Test Plan: N/A

Reviewers: kamil, tomek, ashoat

Reviewed By: ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12428