[web][native] Display both thread and community names in the modal
Summary:
Display both a community and thread names when a link points at a thread. Show just the community name for community links.

https://linear.app/comm/issue/ENG-8417/show-a-thread-name-to-which-a-link-points-in-the-modal

Web
{F2005965}

{F2005966}

Native
{F2005967}

{F2005968}

Long names, e.g. `A thread with a long name containing words with a lot of characters like aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa`:
{F2012226}

{F2012227}

Depends on D12421

Test Plan: Checked on both platforms and for both link types that the names are displayed correctly.

Reviewers: inka, ginsu, ashoat

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D12422