[lib] Define steps in link acceptance hook
Summary:
This makes it easier to reason about the flow and to add a new step.

https://linear.app/comm/issue/ENG-8250/join-a-thread-after-joining-a-community

Depends on D12423

Test Plan:
Tested with the rest of the stack. Checked if joining communities still works. Also, added some timeouts in the auth keyserver effect and in the thread join responder and verified that the loading status was correct.

Note: didn't test the secondary keyserver invite links - they don't work currently because of https://linear.app/comm/issue/ENG-7592/invite-links-of-the-secondary-keyserver-dont-work.

Reviewers: inka, ginsu, kamil

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12424