[web] check if version is supported by identity service
Summary:
call the versionSupported identity client method to check if the client code version is supported. if it's not supported by identity and the client is logged in, log out. regardless of whether or not client is logged in, display modal instructing user to reload/refresh.

Depends on D12406

Test Plan:
- bumped min version supported on staging identity to 1000, confirmed that modal was displayed and i was logged out on web
- modified function returned by useVersionSupportedByIdentity to always throw, confirmed that log appeared in console but user was not logged out and modal was not displayed

Reviewers: ashoat!

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D12407