[web] remove ping component
Summary:
we are now calling the identity service from MinVersionHandler on web, so there's no need for this ping component anymore. the point was to make sure that the identity client is callable via gRPC-web on prod, which is achieved now in MinVersionHandler.

Depends on D12407

Test Plan: made sure there were no references to IdentityPing left in code

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D12408