[lib] Use sqlite for message store local instead of redux persist
Summary:
This diff is in anticipation for when we want to use sqlite for message store local instead of redux persist. This should only be landed after the message store local sqlite diffs have been landed on master for a while and no errors are confirmed with the assertion function.

Depends on D12384

Test Plan: When landing, will verify that the sqlite data is properly reflected on app start on both native and web

Reviewers: tomek, kamil

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12409