[lib] Remove `getCurrentUser`
Summary:
We needed `getCurrentUser` in order to "patch" permissions based on block. However, we're now considering filtered permissions as part of `useThreadsWithPermission` so we can handle blocks as part of permission check vs. `RawThreadInfo` -> `ThreadInfo`. This greatly simplifies deprecation of `memberHasAdminPowers` on the client.

---

Depends on D12154

Test Plan: `flow` + ensuring that `useExistingThreadInfoFinder`, `threadInfoFromRawThreadInfo`, etc. continue to work as expected.

Reviewers: ashoat, ginsu, tomek, varun

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D12158