[keyserver/lib] update farcaster channel tag blob payload to include keyserver id
Summary:
While working on the auto-join communities feature, I discovered an issue with the Farcaster channel tag blob hash method. The current setup allows the download of a Farcaster channel tag blob for a community not part of the keyserver.

The Farcaster channel tag blob hash was composed solely of the farcasterChannelID. This means that if a user in Keyserver A creates a Farcaster channel tag for the /comm Farcaster channel, a user in Keyserver B, who also follows the /comm Farcaster channel, would receive the community info. However, this community would not be part of Keyserver B.

Since communities are localized to their keyserver, we should include the keyserver ID as part of the blob hash to ensure proper download blob functionality in a multikeyserver environment.

Depends on D12062

Test Plan: `flow` + confirmed that the creating + updating + deleting farcaster tag all still worked

Reviewers: tomek, kamil, ashoat

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12063