[keyserver] Remove threadTypes.LOCAL from newThreadRequestInputValidator
Summary:
Currently, `LOCAL` is the same value as `COMMUNITY_SECRET_SUBTHREAD`, so no changes should occur to validation as a result of this diff.

In a later diff, I'll reassign `LOCAL` to a new constant. Going forward, `LOCAL` will be used for thick threads only, which should never be created by a keyserver. So it makes sense to leave this out of `newThreadRequestInputValidator`.

Depends on D12482

Test Plan: Flow

Reviewers: kamil, marcin

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D12483