[lib] Temporarily switch LOCAL to COMMUNITY_SECRET_SUBTHREAD for pending threads
Summary:
These two constants were hardcoded to be the same value. I used `LOCAL` instead of `COMMUNITY_SECRET_SUBTHREAD` in this case as a reminder to the author that we should be created thick threads here instead of thin threads eventually.

I'm ready to introduce a separate constant for `LOCAL`, but we'll need to do some additional work before we're able to switch to using it for new pending threads. For now I'll move this code to continue to use the same constant (`COMMUNITY_SECRET_SUBTHREAD`, which will stay as 4), but I'll return it back to using `threadTypes.LOCAL` once we're ready.

Depends on D12486

Test Plan: Flow

Reviewers: kamil, marcin

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D12487