[lib] Introduce `useThreadFrozenDueToViewersBlock`
Summary:
We introduce a "hook version" of `threadFrozenDueToViewersBlock` which will allow us to set `skipMemberAdminRoleCheck` to `true` in the call to `threadIsWithBlockedUserOnly` in `[web/native]/ChatInputBar` and use the "updated" community role logic instead.

---

Depends on D12265

Test Plan: So far just `flow` and some log statements to ensure that value is as expected + same as before.

Reviewers: ashoat, ginsu, tomek, varun

Reviewed By: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12287