[native] update QR login flow to use `QRAuthContext`
Summary:
[ENG-8385](https://linear.app/comm/issue/ENG-8385/qr-code-login-is-not-working-on-master).

QR code login is not working right now. As described [here](https://linear.app/comm/issue/ENG-8385/qr-code-login-is-not-working-on-master#comment-a8a82224) (Point 3) and [here](https://linear.app/comm/issue/ENG-8385/qr-code-login-is-not-working-on-master#comment-c4e8125c) the issue is occuring because of the fact that `useQRAuth` is unmounting before it can send `SECONDARY_DEVICE_REGISTRATION_SUCCESS`. This stack is refactoring the code to move logic to context (see [here](https://linear.app/comm/issue/ENG-8385/qr-code-login-is-not-working-on-master#comment-647d811a)).

Rendering context and de-duplicating logic.

Depends on D12450

Test Plan: Test QR login flow on native. Adding a bunch of logs to make sure it works and all messages are sent.

Reviewers: bartek