[native] remove `useQRAuth` hook
Summary:
[ENG-8385](https://linear.app/comm/issue/ENG-8385/qr-code-login-is-not-working-on-master).

QR code login is not working right now. As described [here](https://linear.app/comm/issue/ENG-8385/qr-code-login-is-not-working-on-master#comment-a8a82224) (Point 3) and [here](https://linear.app/comm/issue/ENG-8385/qr-code-login-is-not-working-on-master#comment-c4e8125c) the issue is occuring because of the fact that `useQRAuth` is unmounting before it can send `SECONDARY_DEVICE_REGISTRATION_SUCCESS`. This stack is refactoring the code to move logic to context (see [here](https://linear.app/comm/issue/ENG-8385/qr-code-login-is-not-working-on-master#comment-647d811a)).

Moving this code to a context - reduces some complexity.

Depends on D12452

Test Plan: Test QR login flow on native and web. Adding a bunch of logs to make sure it works and all messages are sent.

Reviewers: bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12453