[lib][native] Send backup keys in QR auth only if requested
Summary:
Addresses [[ https://linear.app/comm/issue/ENG-8225/skip-fetching-backup-if-the-secondary-device-being-added-is-a | ENG-8225 ]].
Added a boolean to `SECONDARY_DEVICE_REGISTRATION_SUCCESS` message that determines if backup keys should be sent.
For keyserver, it can be set to false in D12213.

Test Plan: QR login on native and web. Verified that backup keys are sent, unless hardcoding the flag to `false`.

Reviewers: varun, ashoat

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D12327