[backup] Hotfix Blob service client auth
Summary:

Noticed that Backup service had code respoinsible for extracting UserIdentity (auth metadata),
but it was never supplied to `BlobServiceClient` so Backup -> Blob service requests were unauthenticated.

This is not the nicest solution but it works. A proper solution will be introduced in the next diffs.

Test Plan:

Backup service on staging

Reviewers: kamil

Subscribers: